fix(ci): resolve backend dependency issues and frontend linting warnings
