+++ backend/pyproject.toml
@@ -46,6 +46,7 @@ opentelemetry-sdk = ">=1.27.0"
 # --- Utility ---
 httpx = ">=0.28.1,<1.0.0"
 jsonschema = "^4.26.0"
+greenlet = "^3.5.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.0"
