+++ frontend/src/components/theme-provider.tsx
@@ -1,24 +1,11 @@
-import { createContext, useContext, useEffect, useState, type ReactNode } from "react";
-
-export type Theme = "dark" | "light" | "system";
+import { useEffect, useState, type ReactNode } from "react";
+import { ThemeProviderContext, type Theme, type ThemeProviderState } from "./theme-context";
 
 type ThemeProviderProps = {
   children: ReactNode;
   defaultTheme?: Theme;
   storageKey?: string;
 };
-
-type ThemeProviderState = {
-  theme: Theme;
-  setTheme: (theme: Theme) => void;
-};
-
-const initialState: ThemeProviderState = {
-  theme: "system",
-  setTheme: () => null,
-};
-
-const ThemeProviderContext = createContext<ThemeProviderState>(initialState);
 
 export function ThemeProvider({
   children,
@@ -62,11 +49,3 @@ export function ThemeProvider({
 
   return <ThemeProviderContext.Provider value={value}>{children}</ThemeProviderContext.Provider>;
 }
-
-export const useTheme = () => {
-  const context = useContext(ThemeProviderContext);
-  if (context === undefined) {
-    throw new Error("useTheme must be used within a ThemeProvider");
-  }
-  return context;
-};
