+++ frontend/src/components/ui/button.tsx
@@ -59,4 +59,4 @@ function Button({
   );
 }
 
-export { Button, buttonVariants };
+export { Button };
